requirements.txt
asgiref==3.2.10
async-timeout==3.0.1
certifi==2020.6.20
cffi==1.14.4
dj-database-url==0.5.0
Django==3.1.3
django-heroku==0.3.1
django-multiselectfield==0.1.12
django-url-params==1.0.2
gunicorn==20.1.0
hiredis==1.1.0
hyperlink==20.0.1
Js2Py==0.70
Pillow==7.2.0
pipwin==0.5.0
pycparser==2.20
pyjsparser==2.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
regex==2020.9.27
secure-smtplib==0.1.1
service-identity==18.1.0
six==1.15.0
urllib3==1.25.10
virtualenv==20.0.35